put into required format
